refactor the styles to re-use the Row and Cell
